chore: add release please config files and delete changelogs (#436)
## Purpose

This PR adds the necessary files for the Github Release Please workflow.
This includes a config page and a github action (I also deleted the
antiquated changelogs in some apps)

## Approach

Explicitly followed the set-up instructions found
[here](https://github.com/googleapis/release-please/blob/main/docs/manifest-releaser.md).

I needed to open a branch with these changes in order to follow those
steps ^. Ultimately, from use of the CLI I have created
[this](https://github.com/contentful/marketplace-partner-apps/pull/438)
first Release PR.

The manifest file was automatically created by release-please and is in
that other PR ^ which will override this empty manifest file. (Needed
the empty file for it to create the new one).