fix: update isRuleValid to run getValue if it exists on the field (#504)
## Purpose

Flexfields tries to call `.getValue` on a field when `.value` is falsey.
If a field that does not have `.getValue`, but they have an empty field,
it will try to call`.getValue` resulting in an error.

## Approach

Putting a `?` before invocation allows typescript to null-check if the
method exists and then run it. So instead of `.getValue()` it is now
`.getValue?.()`